Change num_bottles to n_bottles, clause to fragment
The num_bottles change is just because it's unnecessarily long but
still an abbreviation. n_bottles conveys the same level of meaning
but is shorter and I just like it a little better.

I decided to change clause to fragment because at first people didn't
understand what I meant by clause, and I think fragment might be
a little more obvious, if a little longer.