Get rid of the `homologous_pairs` function. Use name/arity.
Since `List.zip` is so straightforward, `homologous_pairs` kind of lost
it's self-documenting purpose. I still think `mutation?` is nice.

Using the name/arity method of refercing the function is kinda nice. I
want to get used to using that method for cases where the function takes
more than one param, as it's less verbose and the verbosity is not
helpful in understanding what's going on because the params have no
meaningful names anyway.