Use case and binary pattern matching to simplify things
Allowed me to get rid of a couple of methods dealing with testing to see
if the number had the country code included and to strip the country
code. All of that could definitely be considered part of validation, and
it's clear enough to put it in the validate method, I think.