Use multiple function heads instead of case. Use constants.
I realized that I wasn't using my @us_phone_number_length constant after
that last refactoring, so I re-introduced it, but because it was so long
it was making the lines too long in the multiple function heads version,
so I changed the name to just `@number_length` which I actually like
better because nowhere in the readme does it specify that it's a US
phone number; I just assumed it was because it looks like one. But
perhaps other countries us that same formatting.