Initial attempt at ruby space age, using method_missing.
I was getting test failures in the least significant digits of the ages
for some planets, so I cheated and looked at the example solution to see
what numbers were being used for the calculations. Turns out my errors
were because I was doing the calculation for earth first and using that
to generate the ages on other planets, but since I was rounding the
result for earth, the ages on other planets lost that precision. So, I
took a cue from the example and translated the orbital periods for each
planet into seconds to start off.

Since I looked at the example, which used define_method instead of
method_missing, I realize this version is more complicated than
necessary, but I'm going to submit it anyway because I initially reached
for method_missing, and it's a good learning experience that
define_method is simpler if you already know the list of methods you
will need.