Eliminate the conditional in potential response to a nitpick
I'm not sure that I understood what the nitpicker was getting at, and
I'm not sure if this is better. It elmininates the conditional in
`additions_for_verse` but it adds the extra work of rejecting the empty
additions. Kind of a tradeoff. We'll see what people think of this
iteration.