Remove the is_prime? function and its use
It turns out it wasn't necessary because the algorithm of the Sieve of
Eratosthenes implies that the next item in the candidates list will
always be a prime, so you don't have to test for that. You just add it
to the list of primes and then reject all multiples of that value from
the remaining list of candidates.