Use `Enum.filter_map` to filter matched alergies and map to names
I think this is a nice way of thinking about the problem. You first
filter the allergy tuples that match and then map the tuple to just
the name. This allows the list to be iterated in the order expected
rather than the reverse order.

On the down side, you no longer see the names of the elements of the
tuple or the name of the result in the code, so it might not be quite as
obvious what is going on. I could go either way.