Elminate more duplication and make template read like the song
I totally stole ideas for these changes from other people's awesome
exercises that I was able to read. Specifically the `soMany` and the
`take` functions. I really like how the template reads now, and there's
less duplication. Maybe not quite as easy to read as the previous
iteration, but the previous iteration had quite a bit of duplication.