Use multipleOf instead of divisibleBy, property instead of function
I think I forgot to commit my actual first iteration that I submitted to
exercism. This is my second iteration. The first iteration still used a
Year "class" but the helper function was `divisibleBy` and `isLeapYear`
was a function, but it didn't need to be one, it could just be a
property, calculated once.