Remove function that did not do what it was called
The contains-letters happened to do what I needed it to do, but it did
not do what it says, i.e. return if a message contains letters. It only
worked in the context of if everything was uppercase to begin with.

So, I just moved that logic inline in is-shout, because those tests only
work and make sense together.