Make a few minor improvements to elixir pythagorean triplet
Don't use Enum methods for sum and product, since these are super
straight forward and don't need to handle variable numbers of elements
in the list.

Reduce a couple of triplet options by reducing the max end of the range
as well as the min end of the range slightly.